+++ tests/TestCase.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Sfneal\Helpers\Arrays\Tests;
+
+class TestCase extends \Orchestra\Testbench\TestCase
+{
+}